+++ game/pieces.py
@@ -1,4 +1,5 @@
 from game.exceptions import LimitedMove, InvalidMove, OutOfBoard
+from emoji import emojize
 
 
 class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bishop, Knight, Position = [row , column]
@@ -7,23 +8,24 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
         self.__color__ = color
         self.__position__ = initial_position
         self.__image__ = None
+        self.set_images()
     
     def set_images(self):
         images_white = {
-                        'Queen': 'q',
-                        'King': 'k',
-                        'Pawn' : 'p',
-                        'Knight': 'n',
-                        'Rook': 'r',
-                        'Bishop' : 'b'
+                        'Queen': emojize(':crown:'),
+                        'King': emojize(':elf:'),
+                        'Pawn' : emojize(':nose:'),
+                        'Knight': emojize(':unicorn:'),
+                        'Rook': emojize(':moai:'),
+                        'Bishop' : emojize(':trident_emblem:')
                         }
         images_black = {
-                        'Queen': 'Q',
-                        'King': 'K',
-                        'Pawn' : 'P',
-                        'Knight': 'N',
-                        'Rook': 'R',
-                        'Bishop' : 'B'
+                        'Queen': emojize(':blossom:'),
+                        'King': emojize(':prince:'),
+                        'Pawn' : emojize(':bust_in_silhouette:'),
+                        'Knight': emojize(':horse_face:'),
+                        'Rook': emojize(':office_building:'),
+                        'Bishop' : emojize(':maple_leaf:')
                         }
         if self.__color__ == 'w':
             self.__image__ = images_white[self.__name__]
@@ -33,6 +35,8 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
     def change_position(self, new_position):
         self.__position__ = new_position
 
+    ### Checks if the movement is straight. If the column or row doesnt change,
+    ### its moving in a straight line
     def straight_line(self, new_position):
         row = self.__position__[0]
         column = self.__position__[1]
@@ -45,6 +49,8 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
             squares = self.straight_column(row, column, new_position)
             return squares
     
+    ### Makes a list with each position the piece goes
+    ### to before getting to the new position.
     def straight_row(self, row, column, new_position):
         squares = []
         if column < new_position[1]:
@@ -65,6 +71,8 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
                 squares.append([row - x - 1, column])
         return squares
     
+    ### Checks if the movement is diagonal, both row and column
+    ### should be atleast different. Then calls for check type diagonal 
     def diagonal(self, new_position):
         row = self.__position__[0]
         column = self.__position__[1]
@@ -79,6 +87,9 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
         else:
             return False
     
+    ### Depending on how the row and column changes, it returns a
+    ### string with the name of the function that should be call
+    ### to get the positions
     def check_type_diagonal(self, row, column, new_position):
         if row > new_position[0] and column > new_position[1]:
             return "left_up"
@@ -89,7 +100,8 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
         elif row < new_position[0] and column < new_position[1]:
             return "right_down"
 
-
+    ### Makes a list with each position the piece goes
+    ### to before getting to the new position.
     def right_down(self,row, column, new_position):
         squares = []
         for i in range(new_position[0] - row):
@@ -121,7 +133,9 @@ class Pieces(): ###Color = black or white, Piece = Queen, King, Pawn, Rook, Bish
             column -= 1
             squares.append([row, column])
         return squares
-        
+    
+    ### A movement cant be diagonal and straight or none at the same time
+    ### (only exception would be knight)
     def valid_or_invalid(self, new_position):
         straight = self.straight_line(new_position)
         diagonal = self.diagonal(new_position)
@@ -212,6 +226,16 @@ class Pawn(Pieces):
             return squares
         except Exception as e:
             raise
+    
+    ### A pawn can change to another piece (excluding King),
+    ### when it reaches the other side of the board
+    def change_pawn(self):
+        if self.__position__[0] == 7 and self.__color__ == 'w':
+            return True
+        elif self.__position__[0] == 0 and self.__color__ == 'b':
+            return True
+        else:
+            return False
 
 class Knight(Pieces):
     def limit(self, new_position):